+++ cypress/elements/optionsModal/index.js
@@ -63,7 +63,7 @@ export { setCustomSubtitle } from './subtitle'
 
 export { clickOutliersCheckbox } from './outliers'
 
-export { setItemToAxis } from './series'
+export { setItemToAxis, setItemToType } from './series'
 
 export {
     toggleLegend,
@@ -83,4 +83,5 @@ export {
     expectLegendKeyOptionToBeDisabled,
     expectLegendKeyToBeHidden,
     expectLegendKeyToBeVisible,
+    expectLegedKeyItemAmountToBe,
 } from './legend'