+++ cypress/elements/optionsModal/series.js
@@ -4,3 +4,10 @@ export const setItemToAxis = (itemIndex, axis) =>
         .eq(itemIndex)
         .findBySel(`item-axis-${axis}`)
         .click()
+
+export const setItemToType = (itemIndex, type) =>
+    cy
+        .getBySel('series-table-item')
+        .eq(itemIndex)
+        .findBySel(`item-type-${type}`)
+        .click()