+++ cypress/utils/window.js
@@ -100,6 +100,24 @@ export const expectWindowConfigSeriesToHaveTrendline = expectedTL =>
             expect(actualTL.zIndex).to.eq(expectedTL.zIndex)
         })
 
+export const expectWindowConfigSeriesItemToBeType = (itemIndex, type) =>
+    cy
+        .window()
+        .its(CONFIG_PROP)
+        .its(SERIES_PROP)
+        .then(series => {
+            expect(series[itemIndex].type).to.eq(type)
+        })
+
+export const expectWindowConfigSeriesItemToNotHaveType = itemIndex =>
+    cy
+        .window()
+        .its(CONFIG_PROP)
+        .its(SERIES_PROP)
+        .then(series => {
+            expect(series[itemIndex].type).to.be.undefined
+        })
+
 export const expectWindowConfigYAxisToHaveTitleText = text =>
     cy
         .window()
@@ -195,6 +213,19 @@ export const expectWindowConfigSeriesItemToHaveLegendSet = (
             )
         })
 
+export const expectEachWindowConfigSeriesItemToHaveLegendSet = expectedLS =>
+    cy
+        .window()
+        .its(CONFIG_PROP)
+        .its(SERIES_PROP)
+        .then(series =>
+            series.forEach(seriesItem =>
+                seriesItem.data.every(item =>
+                    expect(item.legendSet).to.eq(expectedLS)
+                )
+            )
+        )
+
 export const expectWindowConfigSeriesItemToNotHaveLegendSet = seriesItemName =>
     cy
         .window()
@@ -204,6 +235,17 @@ export const expectWindowConfigSeriesItemToNotHaveLegendSet = seriesItemName =>
             const seriesItem = series.find(item => item.name === seriesItemName)
             seriesItem.data.every(item => expect(item).to.be.a('number'))
         })
+
+export const expectEachWindowConfigSeriesItemToNotHaveLegendSet = () =>
+    cy
+        .window()
+        .its(CONFIG_PROP)
+        .its(SERIES_PROP)
+        .then(series =>
+            series.forEach(seriesItem =>
+                seriesItem.data.every(item => expect(item).to.be.a('number'))
+            )
+        )
 
 export const expectWindowConfigSeriesDataLabelsToHaveColor = (
     seriesItemIndex,