+++ packages/app/src/components/VisualizationOptions/Options/ColorSet.js
@@ -84,20 +84,33 @@ const ColorSetPreview = ({ colorSet, disabled }) => (
     >
         {colorSet?.patterns &&
             colorSet.patterns.map((pattern, index) => (
-                <div key={`pattern-${index}`}>
-                    <svg
-                        width={14}
-                        height={14}
-                        xmlns="http://www.w3.org/2000/svg"
-                    >
-                        <path
-                            d={pattern.path}
-                            stroke={pattern.color}
-                            strokeWidth={2}
-                            fill="none"
-                        />
-                    </svg>
-                </div>
+                <svg
+                    xmlns="http://www.w3.org/2000/svg"
+                    width={14}
+                    height={14}
+                    key={`pattern-${index}`}
+                >
+                    <defs>
+                        <pattern
+                            id={`bg${index}`}
+                            patternUnits="userSpaceOnUse"
+                            width={pattern.width}
+                            height={pattern.height}
+                        >
+                            <path
+                                d={pattern.path}
+                                stroke={pattern.color}
+                                strokeWidth={2}
+                                fill="none"
+                            />
+                        </pattern>
+                    </defs>
+                    <rect
+                        height="14"
+                        width="14"
+                        fill={`url(#bg${index})`}
+                    ></rect>
+                </svg>
             ))}
         {colorSet?.colors &&
             colorSet.colors.map(color => (